Add adchannel in the main file
