+++ junk.go
@@ -3,8 +3,6 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-
-	"github.com/mitchellh/mapstructure"
 )
 
 // /*Message is a struct*/
@@ -42,14 +40,14 @@ func junk() {
 	}
 }
 
-func addChannel(data interface{}) (Channel, error) {
-	var channel Channel
-	err := mapstructure.Decode(data, &channel)
-	if err != nil {
-		return channel, err
-	}
-	channel.ID = "1"
+// func addChannel(data interface{}) (Channel, error) {
+// 	var channel Channel
+// 	err := mapstructure.Decode(data, &channel)
+// 	if err != nil {
+// 		return channel, err
+// 	}
+// 	channel.ID = "1"
 
-	fmt.Printf("the channel is : %#v\n", channel)
-	return channel, nil
-}
+// 	fmt.Printf("the channel is : %#v\n", channel)
+// 	return channel, nil
+// }
